move bootstrap classes to package com.zendesk.maxwell.bootstrap
